# docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ResponseData_any_": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"items": {},
						"type": "array"
					},
					"data": {
						"nullable": true
					}
				},
				"required": [
					"status",
					"message",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ResponseData_string-or-null_": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"items": {},
						"type": "array"
					},
					"data": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"status",
					"message",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthErrorMessage": {
				"type": "string",
				"enum": [
					"USER_ALREADY_EXISTS",
					"USER_NOT_FOUND",
					"EMAIL_NOT_FOUND",
					"EMAIL_EXIST",
					"OTP_SENT_RECENTLY",
					"OTP_INVALID",
					"TOKEN_REQUIRED",
					"INVALID_TOKEN",
					"INVALID_ACCESS_TOKEN",
					"INVALID_REFRESH_TOKEN",
					"ACCOUNT_INACTIVE",
					"ACCOUNT_NOT_FOUND",
					"PASSWORD_NOT_MATCH",
					"UPDATE_PASSWORD_FAIL",
					"ROLLBACK_FAILED",
					"INTERNAL_SERVER_ERROR",
					"VALIDATION_ERROR",
					"UNKNOWN_ERROR"
				],
				"description": "Union type for all possible Auth error messages"
			},
			"AuthErrorResponse": {
				"description": "Specialized response for Auth errors to show possible messages in Swagger",
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"$ref": "#/components/schemas/AuthErrorMessage",
						"description": "Auth error code. Possible values:\n- USER_ALREADY_EXISTS: User with this phone/email already exists\n- USER_NOT_FOUND: User not found\n- EMAIL_NOT_FOUND: Email not found during OTP request\n- INVALID_OTP: OTP is incorrect or expired\n- PASSWORD_NOT_MATCH: Incorrect password\n- ACCOUNT_INACTIVE: Account is currently blocked or inactive\n- INVALID_TOKEN: Token is invalid or expired\n- TOKEN_REQUIRED: Token is missing\n- EMAIL_EXIST: OTP requested too frequently for this email\n- INTERNAL_SERVER_ERROR: unexpected server error"
					},
					"error": {
						"items": {},
						"type": "array"
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"status",
					"message",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterRequest": {
				"description": "REQUEST DTOs",
				"properties": {
					"phone": {
						"type": "string",
						"description": "Phone number (Vietnam format)",
						"example": "0912345678",
						"pattern": "^(0|84)(3|5|7|8|9)([0-9]{8})$"
					},
					"email": {
						"type": "string",
						"description": "Email address",
						"example": "user@example.com",
						"pattern": "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$"
					},
					"firstName": {
						"type": "string",
						"description": "User first name",
						"minLength": 1,
						"maxLength": 30
					},
					"lastName": {
						"type": "string",
						"description": "User last name",
						"minLength": 1,
						"maxLength": 30
					},
					"birthday": {
						"type": "string",
						"description": "Date of birth (ISO8601 string)",
						"example": "1990-01-01"
					},
					"gender": {
						"type": "string",
						"description": "Gender",
						"example": "Male",
						"pattern": "^(Male|Female)$"
					},
					"password": {
						"type": "string",
						"description": "Password",
						"minLength": 8,
						"pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d).{8,}$"
					},
					"currentCity": {
						"type": "string",
						"maxLength": 100
					},
					"currentDistrict": {
						"type": "string",
						"maxLength": 100
					},
					"currentJob": {
						"type": "string",
						"maxLength": 30
					}
				},
				"required": [
					"phone",
					"email",
					"firstName",
					"lastName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OTPResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"ttl": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseData_OTPResponse_": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"items": {},
						"type": "array"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/OTPResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"status",
					"message",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OTPRequest": {
				"properties": {
					"email": {
						"type": "string",
						"pattern": "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VerifyOTPRequest": {
				"properties": {
					"email": {
						"type": "string",
						"pattern": "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$"
					},
					"otp": {
						"type": "string",
						"minLength": 6,
						"maxLength": 6
					}
				},
				"required": [
					"email",
					"otp"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseData_OTPResponse-or-any_": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"items": {},
						"type": "array"
					},
					"data": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/OTPResponse"
							},
							{}
						],
						"nullable": true
					}
				},
				"required": [
					"status",
					"message",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResetPasswordRequest": {
				"properties": {
					"resetToken": {
						"type": "string"
					},
					"password": {
						"type": "string",
						"description": "New password",
						"minLength": 8,
						"pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d).{8,}$"
					}
				},
				"required": [
					"resetToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginResponse": {
				"description": "RESPONSE DTOs",
				"properties": {
					"accessToken": {
						"type": "string"
					},
					"refreshToken": {
						"type": "string"
					},
					"expiresIn": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"accessToken",
					"refreshToken",
					"expiresIn"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseData_LoginResponse-or-any_": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"items": {},
						"type": "array"
					},
					"data": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/LoginResponse"
							},
							{}
						],
						"nullable": true
					}
				},
				"required": [
					"status",
					"message",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginRequest": {
				"properties": {
					"identifier": {
						"type": "string",
						"description": "Phone or Email"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"identifier"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RefreshTokenResponse": {
				"properties": {
					"accessToken": {
						"type": "string"
					},
					"refreshToken": {
						"type": "string"
					}
				},
				"required": [
					"accessToken",
					"refreshToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseData_RefreshTokenResponse-or-any_": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"items": {},
						"type": "array"
					},
					"data": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/RefreshTokenResponse"
							},
							{}
						],
						"nullable": true
					}
				},
				"required": [
					"status",
					"message",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RefreshTokenRequest": {
				"properties": {
					"refreshToken": {
						"type": "string"
					}
				},
				"required": [
					"refreshToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseData_string-or-any_": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"items": {},
						"type": "array"
					},
					"data": {
						"anyOf": [
							{
								"type": "string"
							},
							{}
						],
						"nullable": true
					}
				},
				"required": [
					"status",
					"message",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LogoutRequest": {
				"properties": {
					"token": {
						"type": "string"
					}
				},
				"required": [
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ChangePasswordRequest": {
				"properties": {
					"oldPassword": {
						"type": "string"
					},
					"newPassword": {
						"type": "string",
						"description": "New password",
						"minLength": 8,
						"pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d).{8,}$"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "tro-tot-vn-be",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "lamdx4"
		}
	},
	"paths": {
...
		"/auth/register": {
			"post": {
				"operationId": "RegisterAccount",
				"responses": {
					"201": {
						"description": "User registered successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_string-or-null_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponse"
								}
							}
						}
					},
					"409": {
						"description": "Conflict",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponse"
								}
							}
						}
					}
				},
				"description": "Register a new account",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterRequest"
							}
						}
					}
				}
			}
		},
		"/auth/send-otp-register": {
			"post": {
				"operationId": "SendOTPRegister",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_OTPResponse_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponse"
								}
							}
						}
					},
					"429": {
						"description": "Email Exists / Rate Limit",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponse"
								}
							}
						}
					}
				},
				"description": "Send OTP for registration",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/OTPRequest"
							}
						}
					}
				}
			}
		},
		"/auth/verify-otp-register": {
			"post": {
				"operationId": "VerifyOTPRegister",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponse"
								}
							}
						}
					}
				},
				"description": "Verify OTP for registration",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/VerifyOTPRequest"
							}
						}
					}
				}
			}
		},
		"/auth/forgot-password": {
			"post": {
				"operationId": "ForgotPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_OTPResponse-or-any_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponse"
								}
							}
						}
					},
					"429": {
						"description": "Email exists / Rate Limit",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponse"
								}
							}
						}
					}
				},
				"description": "Request password reset OTP",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/OTPRequest"
							}
						}
					}
				}
			}
		},
		"/auth/verify-otp": {
			"post": {
				"operationId": "VerifyOtp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponse"
								}
							}
						}
					}
				},
				"description": "Verify password reset OTP",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/VerifyOTPRequest"
							}
						}
					}
				}
			}
		},
		"/auth/reset-password": {
			"post": {
				"operationId": "ResetPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponse"
								}
							}
						}
					}
				},
				"description": "Reset password with reset token",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ResetPasswordRequest"
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_LoginResponse-or-any_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponse"
								}
							}
						}
					},
					"423": {
						"description": "Account Inactive",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponse"
								}
							}
						}
					}
				},
				"description": "Login to account",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginRequest"
							}
						}
					}
				}
			}
		},
		"/auth/refresh-token": {
			"post": {
				"operationId": "RefreshToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_RefreshTokenResponse-or-any_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponse"
								}
							}
						}
					}
				},
				"description": "Refresh access token",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RefreshTokenRequest"
							}
						}
					}
				}
			}
		},
		"/auth/logout": {
			"post": {
				"operationId": "Logout",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_string-or-any_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponse"
								}
							}
						}
					}
				},
				"description": "Logout from account",
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LogoutRequest"
							}
						}
					}
				}
			}
		},
		"/auth/change-password": {
			"post": {
				"operationId": "ChangePassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_string-or-any_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponse"
								}
							}
						}
					}
				},
				"description": "Change account password",
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChangePasswordRequest"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}